Add specialisation machinery for trivial traversal
[Auto-deref specialisation] is introduced to type library traversals,
the public API of which is via the `TriviallyTraverses` trait and the
`noop_if_trivially_traversable` macro.  Traversals invoked via the macro
will then be no-ops if the interner "trivially traverses" the type being
traversed *without requiring that type to implement the relevant
traversable trait*.

The derive macros for the traversable traits are also updated to use the
`noop_if_trivially_traversable` macro when traversing (unskipped) fields
that are neither generic nor guaranteed trivial, which negates the need
for `#[skip_traversal(because_trivial)]` attributes when such cases are
in fact trivial.

[Auto-deref specialisation]: http://lukaskalbertodt.github.io/2019/12/05/generalized-autoref-based-specialization.html