Use rustc_type_ir directly in derived traversables
A generic marker trait, `rustc_type_ir::TriviallyTraverses<T>`, is added
(by which interners can be marked as trivially traversing particular
types).  A generic implementation thereof is then provided for the
`TyCtxt` interner to indicate that it "trivially traverses" all
`TriviallyTraversable` types.  This indirection is necessary because
auto-traits are unstable, whereas the type library is intended to be
stabilised.