Enable skipping of derived traversals, with reason
A new trait, `rustc_middle::ty::TriviallyTraversable`, is automatically
implemented on all types that do not contain anything that may be of
interest to `TyCtxt` traversers.

The derive macros for the traversable traits are also updated to support
a `#[skip_traversal]` helper attribute on potentially non-trivial
fields, the soundness of which is ensured by constraining generated
implementations so that they apply only where the types of such
annotated fields implement the above `TriviallyTraversable` auto-trait.